Complete H1 tag interface
Add an option to disable storing of padding. This will zero out these fields like the MCC tools.

Check that result is none before moving on as there has been instances of result being 0 or 1 and making the code behave in strange ways.

Add support support for unsigned fields

Store negative zeros as strings to preserve them for hash reasons

Change default settings so that checksums are generated, radians are converted to degrees for the json and back to radians for writing, and preservation of straings and padding disabled. Padding will be zeroed out and strings will be stored as literal instead of byte data.