Additional features for Halo 2 tag versioning
Disable preserve version. Tags are upgraded to the latest fieldsets that are available. Bitmaps seem to be broken right now so would avoid that for the time being. This is for testing.

Add an H1Latest enum in the engine tags for versioning. Mostly to mirror the H2 options. Doubt there will ever be more H1 engine tags.

Clean up get_field args to remove pointless args

Add return_size arg for get_fields. This makes the function just return the size of the fieldset we are looking at currently.

Change read_file and write_file to take the engine tag instead and use that to set the file_endian. Endianess can still be overwritten with the file_endian_overwrite arg