Quick fixes
Add a new setting to preserve the size of the field. If it's set to true then we use the size told to use by the existing struct header if it's there. If it's false then we find out what the size of the field is and use that instead.

Fix issue in render_model post process. Seems to be working properly now.