More fixes for the toolset
Make some fields in H1 match the defined names in Guerilla. This is so that the definitions line up with H2

Add code to check if the file given is a proper Halo tag

Make data and tag reference types get their required keys instead of assuming it will unpack into the given length.

Add a argument to generate_tag_dictonary to prepare the json dict for Blender. All this does at the moment is convert floats that were written as strings to preserve signs back into floats

Fix formatting issue in find function in H2 postprocessing